fix: remove InteractionManager from InputFocus - PR introducing it - https://github.com/Expensify/App/pull/60073
