+++ src/libs/actions/InputFocus/index.native.ts
@@ -1,0 +1,8 @@
+import type {Modal} from '@src/types/onyx';
+
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+function inputFocusChange(focus: boolean) {}
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+function composerFocusKeepFocusOn(ref: HTMLElement, isFocused: boolean, modal: Modal, onyxFocused: boolean) {}
+
+export {composerFocusKeepFocusOn, inputFocusChange};
